% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfe.R
\name{mdfa.cfe}
\alias{mdfa.cfe}
\title{Common factor score identification in matrix decomposition factor analysis (MDFA)
; Clustering version}
\usage{
mdfa.cfe(X, r, k, lambda, eps = 1e-06, itemax = 2000)
}
\arguments{
\item{X}{A matrix. n * p data matrix}

\item{r}{An integer. Number of common factors.}

\item{k}{An integer. Number of clusters.}

\item{lambda}{A numeric value. Tuning parameter > 0}

\item{eps}{A numeric value. Convergence criterion.}

\item{itemax}{An integer. Maximum number of iteration}
}
\value{
List of the following objects.\cr
F: A matrix. Common factor score.\cr
Fr: A matrix. Common factor score; rotated.\cr
U: A matrix. Unique factor score.\cr
A: A matrix. Factor loadings.\cr
Ar: A matrix. Factor loadings; rotated.\cr
C: A matrix. Cluster centroid.\cr
Cr: A matrix. Cluster centroid; rotated.\cr
M: A matrix. Membership.\cr
D: A matrix. Square root of uniqueness.\cr
int.factor.corr: A matrix. Inter-factor correlation.\cr
fitted.cov: A matrix. Covariance matrix fitted to X.\cr
history: A numeric vector. Iteration history.\cr
AIC: A numeric value. AIC value.\cr
}
\description{
Specifies common/unique factor scores in MDFA by solving the following minimization problem
min. ||X - (FA' + UD)||^2 + lambda||MC - F||^2 over F, A, U, D, M, and C.\cr
\cr
n: number of observations\cr
p: number of observed variables\cr
r: number of common factors\cr
k: number of clusters\cr
X: n * p data matrix\cr
F: n * r common factor score matrix\cr
U: n * p unique factor score matrix\cr
A: p * r factor loading matrix\cr
D: diagonal matrix of square root of uniqueness\cr
M: n * k membership matrix\cr
C: k * p cluster centroid matrix\cr
lambda : tuning parameter\cr
}
\examples{
set.seed(25258800)
options(warn = -1)
data(jobImage)
X <- scale(as.matrix(jobImage[,-1]))
X <- X[,c(2,3,10,7,8,12)]#Useful, Good, Firm, Quick, Noisy, Busy
lambda <- 0.01
res.list <- list()
f.list <- c()
n.starts <- 250
for(i in 1:n.starts){
  res.list[[i]] <- mdfa.cfe(X, 2, 3, lambda)
  f.list[i] <- min(res.list[[i]]$history)
}
res.cfe <- res.list[[which.min(f.list)]]
plot(res.cfe$Fr)
M <- res.cfe$M #membership
rownames(M) <- jobImage[,1]
M
data.frame(F1 = res.cfe$Ar[,1],
           F2 = res.cfe$Ar[,2],
           uniqueness = diag(res.cfe$D^2))
}
\author{
Naoto Yamashita, \email{nyam@kansai-u.ac.jp}
}
