% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfe.R
\name{mdfa.cfe}
\alias{mdfa.cfe}
\title{Common factor score identification in matrix decomposition factor analysis (MDFA)
; Clustering version}
\usage{
mdfa.cfe(X, r, k, lambda, eps = 1e-06, itemax = 2000)
}
\arguments{
\item{X}{A matrix. n * p data matrix}

\item{r}{An integer. Number of common factors.}

\item{k}{An integer. Number of clusters.}

\item{lambda}{A numeric value. Tuning parameter > 0}

\item{eps}{A numeric value. Convergence criterion.}

\item{itemax}{An integer. Maximum number of iteration}
}
\value{
List of the following objects.\cr
F: A matrix. Common factor score.\cr
Fr: A matrix. Common factor score; rotated.\cr
U: A matrix. Unique factor score.\cr
A: A matrix. Factor loadings.\cr
Ar: A matrix. Factor loadings; rotated.\cr
C: A matrix. Cluster centroid.\cr
Cr: A matrix. Cluster centroid; rotated.\cr
M: A matrix. Membership.\cr
D: A matrix. Square root of uniqueness.\cr
int.factor.corr: A matrix. Inter-factor correlation.\cr
fitted.cov: A matrix. Covariance matrix fitted to X.\cr
history: A numeric vector. Iteration history.\cr
AIC: A numeric value. AIC value.\cr
}
\description{
Specifies common/unique factor scores in MDFA by solving the following minimization problem
min. ||X - (FA' + UD)||^2 + lambda||MC - F||^2 over F, A, U, D, M, and C.\cr
\cr
n: number of observations\cr
p: number of observed variables\cr
r: number of common factors\cr
k: number of clusters\cr
X: n * p data matrix\cr
F: n * r common factor score matrix\cr
U: n * p unique factor score matrix\cr
A: p * r factor loading matrix\cr
D: diagonal matrix of square root of uniqueness\cr
M: n * k membership matrix\cr
C: k * p cluster centroid matrix\cr
lambda : tuning parameter\cr
}
\examples{
X <- read.csv("jobImage.csv")
X <- scale(as.matrix(X[,-1])) #Useful, Good, Firm, Quick, Noisy, Busy
X <- X[,c(2,3,10,7,8,12)]
aics <- c()
lambdas <- 10^(seq(-3, 1, by = 0.5))
res.list <- list()
idx <- 1
n.starts <- 3
for(lambda in lambdas){
  res.list.tmp <- list()
  f.list.tmp <- c()
  for(i in 1:n.starts){
    res.list.tmp[[i]] <- mdfa.cfe(X, 2, 3, lambda)
    f.list.tmp[i] <- min(res.list.tmp[[i]]$history)
  }
  res.list[[idx]] <- res.list.tmp[[which.min(f.list.tmp)]]
  aics[idx] <- res.list[[idx]]$AIC
  idx <- idx + 1
}
plot(res.list[[which.min(aics)]]$Fr)
res.list[[which.min(aics)]]$M
}
\author{
Naoto Yamashita, \email{nyam@kansai-u.ac.jp}
}
