% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfe.R
\name{mdfa.rfe}
\alias{mdfa.rfe}
\title{Common factor score identification in matrix decomposition factor analysis (MDFA)
; Regression version}
\usage{
mdfa.rfe(X, Y, r, lambda, eps = 1e-06, itemax = 2000)
}
\arguments{
\item{X}{A matrix. n * p data matrix}

\item{Y}{A matrix. n * q data matrix of external criterion}

\item{r}{An integer. Number of common factors.}

\item{lambda}{A numeric value. Tuning parameter > 0}

\item{eps}{A numeric value. Convergence criterion.}

\item{itemax}{An integer. Maximum number of iteration}
}
\value{
List of the following objects.\cr
F: A matrix. Common factor score.\cr
Fr: A matrix. Common factor score; rotated.\cr
U: A matrix. Unique factor score.\cr
A: A matrix. Factor loadings.\cr
Ar: A matrix. Factor loadings; rotated.\cr
W: A matrix. Regression coefficients\cr
D: A matrix. Square root of uniqueness.\cr
int.factor.corr: A matrix. Inter-factor correlation.\cr
fitted.cov: A matrix. Covariance matrix fitted to X.\cr
history: A numeric vector. Iteration history.\cr
AIC: A numeric value. AIC value.\cr
}
\description{
Specifies common/unique factor scores in MDFA by solving the following minimization problem
min. ||X - (FA' + UD)||^2 + lambda||Y - FW||^2 over F, A, U, D, and W.\cr
\cr
n: number of observations\cr
p: number of observed variables\cr
q: number of external criterion\cr
r: number of common factors\cr
X: n * p data matrix\cr
F: n * r common factor score matrix\cr
U: n * p unique factor score matrix\cr
A: p * r factor loading matrix\cr
D: diagonal matrix of square root of uniqueness\cr
W: r * q regression coefficient matrix\cr
lambda : tuning parameter\cr
}
\examples{
set.seed(123456)
options(warn = -1)
library(lavaan)
data("HolzingerSwineford1939")
X <- scale(as.matrix(HolzingerSwineford1939[,7:15]))
Y <- matrix(0, nrow(X), 4)
Y[,1] <- ifelse(HolzingerSwineford1939$sex == 1, 1, 0) #sex
Y[,2] <- ifelse(HolzingerSwineford1939$sex == 2, 1, 0)
Y[,3] <- ifelse(HolzingerSwineford1939$school == "Pasteur", 1, 0)
Y[,4] <- ifelse(HolzingerSwineford1939$school == "Grant-White", 1, 0)
aics <- c()
lambdas <- 10^(seq(-3, 1, by = 0.5))
res.list <- list()
idx <- 1
n.starts <- 10
for(lambda in lambdas){
  res.list.tmp <- list()
  f.list.tmp <- c()
  for(i in 1:n.starts){
    res.list.tmp[[i]] <- mdfa.rfe(X, Y, 3, lambda)
    f.list.tmp[i] <- min(res.list.tmp[[i]]$history)
  }
  res.list[[idx]] <- res.list.tmp[[which.min(f.list.tmp)]]
  aics[idx] <- res.list[[idx]]$AIC
  idx <- idx + 1
}
head(res.list[[which.min(aics)]]$Fr)
}
\author{
Naoto Yamashita, \email{nyam@kansai-u.ac.jp}
}
